Update Table of Contents in README.md
Update the existing Table of Contents with: a Table of Contents heading, link to the FAQ, link to the Requested Distributions, and link to License. Additionally, follow the convention of having an `<a name=""></a>` tag for the Table of Contents links, and added tags to each of the listed sections.